Revert "Add redux"
