+++ frontend/package.json
@@ -3,7 +3,6 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@reduxjs/toolkit": "^2.7.0",
     "@testing-library/dom": "^10.4.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.2.0",
@@ -14,7 +13,6 @@
     "react-dom": "^19.1.0",
     "react-google-recaptcha": "^3.1.0",
     "react-markdown": "^10.1.0",
-    "react-redux": "^9.2.0",
     "react-router-dom": "^7.4.1",
     "react-scripts": "^5.0.1",
     "react-syntax-highlighter": "^15.6.1",
