+++ frontend/src/components/Register.js
@@ -1,13 +1,11 @@
 import React, { useState } from 'react';
-import { useDispatch } from 'react-redux';
-import { setLoading, openModal, closeModal } from '../../redux/slices/authSlice';
 import ReCAPTCHA from 'react-google-recaptcha';
-import API from '../../services/api';
-import '../../styles/Auth.css';
+import API from '../services/api';
+import '../styles/Auth.css';
 
 const RECAPTCHA_SITE_KEY = process.env.REACT_APP_RECAPTCHA_SITE_KEY;
 
-const Register = () => {
+const Register = ({ onSwitchToLogin, setLoading }) => {
   const [username, setUsername] = useState('');
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
@@ -16,19 +14,21 @@ const Register = () => {
   const [message, setMessage] = useState('');
   const [recaptchaToken, setRecaptchaToken] = useState('');
   const [registered, setRegistered] = useState(false);
-  const dispatch = useDispatch();
 
   const handleRegister = async (e) => {
     e.preventDefault();
+
     if (password !== confirmPassword) {
       setMessage({ text: 'Passwords do not match.', type: 'error' });
       return;
     }
+
     if (!recaptchaToken) {
       setMessage({ text: 'Please complete the reCAPTCHA.', type: 'error' });
       return;
     }
-    dispatch(setLoading(true));
+
+    setLoading(true);
     try {
       await API.post('/register', { username, email, password, recaptchaToken });
       setRegistered(true);
@@ -39,7 +39,7 @@ const Register = () => {
     } catch (error) {
       setMessage({ text: 'Registration failed. Please try again.', type: 'error' });
     } finally {
-      dispatch(setLoading(false));
+      setLoading(false);
     }
   };
 
@@ -48,7 +48,7 @@ const Register = () => {
       <div className="auth-form">
         <h1>Registration Complete</h1>
         <p className="success-message">{message.text}</p>
-        <button onClick={() => dispatch(openModal('login'))}>Go to Login</button>
+        <button onClick={onSwitchToLogin}>Go to Login</button>
       </div>
     );
   }
@@ -80,7 +80,6 @@ const Register = () => {
         />
         <button
           type="button"
-          className="toggle-password-btn"
           onClick={() => setShowPassword(!showPassword)}
         >
           {showPassword ? '👁️' : '🙈'}
@@ -95,7 +94,7 @@ const Register = () => {
           required
         />
         <button
-          className="toggle-password-btn"
+          className="show-password-btn"
           type="button"
           onClick={() => setShowPassword(!showPassword)}
         >
@@ -108,7 +107,7 @@ const Register = () => {
           onChange={token => setRecaptchaToken(token)}
         />
       </div>
-      <button type="submit" className="auth-btn">Register</button>
+      <button type="submit">Register</button>
       {message && (
         <p className={message.type === 'success' ? 'success-message' : 'error-message'}>
           {message.text}
@@ -116,7 +115,7 @@ const Register = () => {
       )}
       <p className="switch-auth">
         Already have an account?{' '}
-        <button type="button" className='switch-auth-btn' onClick={() => dispatch(openModal('login'))}>
+        <button type="button" onClick={onSwitchToLogin}>
           Login here
         </button>
       </p>
