--- frontend/src/components/layout/Modal.js
@@ -1,15 +1,0 @@
-import React from 'react';
-
-const Modal = ({ isOpen, onClose, children }) => {
-  if (!isOpen) return null;
-  return (
-    <div className="modal-overlay" onClick={onClose}>
-      <div className="modal-content" onClick={e => e.stopPropagation()}>
-        <button className="modal-close" onClick={onClose}>Close Modal</button>
-        {children}
-      </div>
-    </div>
-  );
-};
-
-export default Modal;
