+++ frontend/src/index.css
@@ -39,15 +39,13 @@ html, body {
   background-color: #000; /* Ensure the navbar background is black */
   color: white; /* White text for contrast */
 }
-
-/* Dim and disable the navbar when modal is open */
-.navbar-disabled {
-  pointer-events: none;
-  opacity: 0.5;
-  filter: blur(1px);
-  user-select: none;
+.navbar-brand {
+  font-family: monospace;
+  color: var(--computerAnythingPrimary);
 }
-
+.navbar-brand:hover {
+  color: var(--computerAnythingSecondary);
+}
 .footer {
   background-color: #000; /* Black background */
   color: white; /* White text */
@@ -57,42 +55,3 @@ html, body {
   margin-top: auto; /* Push the footer to the bottom */
 }
 
-/* Style for success messages */
-.success-message {
-  color: green !important;
-  background: #d4edda;
-  border: 1px solid #b7e0c3;
-  border-radius: 4px;
-  font-size: 0.95rem;
-  margin-top: 5px;
-  padding: 8px 12px;
-  text-align: center;
-}
-
-/* Style for error messages */
-.error-message {
-  color: red;
-  background: #d4edda;
-  border: 1px solid #b7e0c3;
-  border-radius: 4px;
-  font-size: 0.95rem;
-  margin-top: 5px;
-  padding: 8px 12px;
-  text-align: center;
-}
-
-.modal-close {
-  margin-bottom: 5px;
-  margin-left: auto;
-  margin-right: auto;
-  background-color: var(--computerAnythingPrimary);
-  color: black;
-  border-radius: 4px;
-}
-
-.modal-close:hover {
-  background-color: black;
-  color: var(--computerAnythingPrimary);
-  cursor: pointer;
-  transition: all 0.3s ease;
-}
