+++ frontend/src/styles/Auth.css
@@ -23,45 +23,19 @@
   font-size: 1rem;
 }
 
-/* BUTTONS: use only class selectors */
-.auth-btn {
+.auth-form button[type="submit"] {
   width: 100%;
   padding: 10px;
-  background: #222;
-  color: #00ff00;
-  border: 2.5px solid #00ff00;
+  background: #007bff;
+  color: white;
+  border: none;
   border-radius: 4px;
   font-size: 1rem;
   cursor: pointer;
-  margin-bottom: 0.5em;
-  box-sizing: border-box;
 }
 
-.auth-btn:hover {
-  background: #00ff00;
-  color: #222;
-  border: 2.5px solid #222;
-  transition: background-color 0.3s ease, color 0.3s ease;
-}
-
-.guest-btn {
-  width: 100%;
-  margin-top: 1em;
-  background: #222;
-  color: #00ff00;
-  border: 2.5px solid #00ff00;
-  border-radius: 4px;
-  font-size: 1rem;
-  cursor: pointer;
-  margin-bottom: 0.5em;
-  box-sizing: border-box;
-}
-
-.guest-btn:hover {
-  background: #00ff00;
-  color: #222;
-  border: 2.5px solid #222;
-  transition: background-color 0.3s ease, color 0.3s ease;
+.auth-form button[type="submit"]:hover {
+  background: #0056b3;
 }
 
 /* Container for input fields with toggle buttons */
@@ -69,30 +43,44 @@
   display: flex;
   align-items: center;
   width: 100%;
-  margin-bottom: 15px;
+  margin-bottom: 15px; /* Add spacing between input fields */
 }
 
 /* Input fields */
 .input-container input {
-  flex: 1;
-  padding-right: 10px;
+  flex: 1; /* Take up all available space */
+  padding-right: 10px; /* Add some space between the input and the emoji */
   border: 1px solid #ccc;
   border-radius: 4px;
   font-size: 1rem;
 }
 
 /* Toggle button for password visibility */
-.toggle-password-btn {
+.input-container button {
   background: none;
   border: none;
   cursor: pointer;
   font-size: 1.2rem;
   color: #555;
-  margin-left: 5px;
+  margin-left: 5px; /* Add spacing between the input and the button */
 }
 
-.toggle-password-btn:hover {
+.input-container button:hover {
   color: #000;
+}
+
+/* Style for success messages */
+.auth-form .success-message {
+  color: green;
+  font-size: 0.9rem;
+  margin-top: 5px;
+}
+
+/* Style for error messages */
+.auth-form .error-message {
+  color: red;
+  font-size: 0.9rem;
+  margin-top: 5px;
 }
 
 /* Switch auth link */
@@ -103,7 +91,7 @@
   text-align: center;
 }
 
-.switch-auth-btn {
+.switch-auth button {
   background: none;
   border: none;
   color: #007bff;
@@ -112,7 +100,7 @@
   padding: 0;
 }
 
-.switch-auth-btn:hover {
+.switch-auth button:hover {
   color: #0056b3;
 }
 
@@ -121,44 +109,8 @@
   .auth-form {
     padding: 15px;
   }
-  .auth-btn,
-  .guest-btn {
+
+  .auth-form button {
     font-size: 0.9em;
   }
-}
-
-/* Loading Page */
-.loading-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  background: rgba(30, 30, 30, 0.7);
-  z-index: 2000;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.loading-spinner {
-  border: 6px solid #f3f3f3;
-  border-top: 6px solid #007bff;
-  border-radius: 50%;
-  width: 48px;
-  height: 48px;
-  animation: spin 1s linear infinite;
-  margin-bottom: 16px;
-}
-
-@keyframes spin {
-  0% { transform: rotate(0deg);}
-  100% { transform: rotate(360deg);}
-}
-
-.loading-text {
-  color: #fff;
-  font-size: 1.2rem;
-  font-weight: bold;
 }
