+++ frontend/src/styles/CreateEditPost.css
@@ -1,3 +1,8 @@
+:root {
+  --computerAnythingPrimary: #02c43c; /* Green */
+  --computerAnythingSecondary: #0250c4; /* Blue */
+}
+
 /* Ensure the entire page takes up the full height */
 html, body {
   height: 100%; /* Full height of the viewport */
@@ -7,18 +12,16 @@ html, body {
 }
 
 /* Container for the create/edit post page */
-/* ...existing code... */
-
 .create-edit-post-container {
-  flex: 1;
+  flex: 1; /* Allow the container to grow and fill the space between the header and footer */
   display: flex;
   flex-direction: row;
   justify-content: space-between;
   padding: 20px;
-  background-color: var(--computerAnythingSecondary);
-  color: var(--computerAnythingPrimary);
+  background-color: var(--computerAnythingSecondary); /* Blue background */
+  color: var(--computerAnythingPrimary); /* Green text */
   font-family: Arial, sans-serif;
-  gap: 20px;
+  gap: 20px; /* Add spacing between the form and preview sections */
 }
 
 /* Form section styling */
@@ -131,7 +134,7 @@ html, body {
 }
 
 /* Back to Blog button styling */
-.back-to-blog-btn, .create-post-submit-btn {
+.back-to-blog-btn {
   width: 100%;
   padding: 10px;
   background-color: var(--computerAnythingPrimary); /* Green background */
@@ -144,7 +147,7 @@ html, body {
   transition: background-color 0.3s ease;
 }
 
-.back-to-blog-btn:hover, .create-post-submit-btn:hover {
+.back-to-blog-btn:hover {
   background-color: #008f00; /* Darker green on hover */
   color: #fff; /* White text on hover */
 }
@@ -181,10 +184,9 @@ html, body {
 
 .form-section .delete-post-btn,
 .delete-post-btn {
-  width: 100%; /* Match the Save Changes button width */
   margin-top: 10px;
-  background-color: #ff0000 !important;
-  color: #fff !important;
+  background-color: #ff0000 !important; /* Red background */
+  color: #fff !important; /* White text */
   border: none;
   border-radius: 5px;
   padding: 10px;
@@ -198,24 +200,4 @@ html, body {
 .delete-post-btn:hover {
   background-color: #cc0000 !important; /* Darker red on hover */
   color: #fff !important; /* White text on hover */
-}
-
-
-@media (max-width: 900px) {
-  .create-edit-post-container {
-    flex-direction: column;
-    align-items: center;
-    padding: 10px;
-    column-gap: 20px;
-    row-gap: 0px;
-  }
-  .form-section,
-  .preview-section {
-    width: 100%;
-    margin: 0;
-    box-sizing: border-box;
-  }
-  .form-section {
-    margin-bottom: 20px;
-  }
 }
