+++ frontend/src/styles/Modal.css
@@ -1,0 +1,41 @@
+.modal-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.5);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 1000;
+}
+
+.modal-content {
+  background: white;
+  padding: 20px;
+  border-radius: 8px;
+  width: 90%;
+  max-width: 500px;
+  position: relative;
+}
+
+.modal-close {
+  position: absolute;
+  top: 10px;
+  right: 10px;
+  background: none;
+  border: none;
+  font-size: 1.5rem;
+  cursor: pointer;
+}
+
+/* Disable pointer events for the background when modal is open */
+.modal-open {
+  pointer-events: none;
+  overflow: hidden; /* Prevent scrolling */
+}
+
+.modal-open .modal-overlay {
+  pointer-events: auto; /* Allow interaction with the modal */
+}
